requirements.txt
clip==1.0
diffusers==0.34.0
huggingface_hub==0.35.3
imageio==2.37.0
matplotlib==3.10.7
numpy==2.3.3
opencv_python==4.12.0.88
pandas==2.3.3
Pillow==11.3.0
pytorch_lightning==2.5.4
Requests==2.32.5
scipy==1.16.2
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.56.1
webdataset==1.0.2